Fixed projectile pushing characters and
Continued the correction on the unreal FP template:
Projectiles no longer pushing characters (desactivated collision for pawn, but kept it on object simulating physics by adding impulse on hit).
Added Enum on Gun to trace owner.
Added Enum on Projectile to trace owner.
Added switch behavior on Projectiles overlapping Player  depending on Projectile owner.

Note: There is no easy way (meaning without creating a custom method) to add custom parameters to the SpawnActor() method. Thus, when the Player spawns a projectile it directcly overlaps with it and fire OnBeginOverlap() before being able to set ProjectileOwner to "Player". This is handle by adding in ABallProjectile::BallProjectile() a default owner "None" and handling it in OnBeginOverlap().