Started FP and TP views for player
Added FP and TP handling via PlayerCharacter class.
Put damage making in cpp in the Projectile Class. Created BallDamageType class for this purpose.

To fix : TP spawns on FP and collides and is visibile by FP.
To fix : damage handling on player not happenning because of Destroy projectiles on capsule overlap thus hit never happens.