Important Maj add of crouch height for FP and projectile spawn fix and Gun Hand IK properly set added coherence animations for gameplay
Added crouch reducing height for FP.
Little refactoring in gun.
Moved the TP spawn point in PC so that it doesn't collide with projectiles spawned by FP.
Solved the issue in which projectiles spawned by TP won't take the flag of their owner.
Simplification of the gun grippoint calcul process.
Rework of the IK hand process.
Fixed a glitch in which one could shoot then run and keep on shootingin this process.
Added mesh rotation when jogging in FP to show non affordance.